Made separate module for general parts of code
