+++ asocial/options/utilities.js
@@ -1,5 +1,3 @@
-/* exported inherit */
-
 'use strict';
 
 /**
