+++ webpack.config.js
@@ -1,0 +1,12 @@
+'use strict';
+
+module.exports = {
+    entry: {
+        vk: './asocial/src/content/networks/vk.js',
+        fb: './asocial/src/content/networks/facebook.js'
+    },
+
+    output: {
+        filename: 'asocial/content/[name].js'
+    }
+};
